feat(IconButton): export IconButton
Expose the earlier internal component `IconButton` for consumers usage
